requirements.txt
streamlit
langchain-google-genai
langchain-openai
langchain-pinecone
pinecone-client
langgraph
langchain-community
langchain-core
langchain-text-splitters
python-dotenv
psycopg
tqdm